TGP-2895: move new UKIMS pages to separate folders
