+++ app/pages/newUkims/NewUkimsNumberPage.scala
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package pages
+package pages.newUkims
 
+import pages.QuestionPage
 import play.api.libs.json.JsPath
 import utils.Constants.ukimsNumberKey
