+++ app/views/newUkims/CyaNewUkimsNumberView.scala.html
@@ -35,7 +35,7 @@
 
     @govukSummaryList(list)
 
-    @formHelper(action = controllers.profile.routes.CyaNewUkimsNumberController.onSubmit()) {
+    @formHelper(action = controllers.newUkims.routes.CyaNewUkimsNumberController.onSubmit()) {
         @govukButton(ButtonViewModel(messages("site.save.continue")))
     }
 }
